Refactor document processing structure and implement new services
- Removed the AWS Textract integration documentation from backend/docs/aws-textract-integration.md as it is no longer relevant.
- Updated import paths in backend/src/app.module.ts and backend/src/app.module.spec.ts to reflect the new directory structure for document processing services.
- Introduced backend/src/document-processor/document-processor.module.ts to encapsulate document processing logic, including AWS Textract and Bedrock services.
- Added backend/src/document-processor/controllers/document-processor.controller.ts to handle document uploads and processing requests.
- Implemented backend/src/document-processor/services/aws-textract.service.ts and backend/src/document-processor/services/aws-bedrock.service.ts for text extraction and medical analysis, respectively.
- Enhanced unit tests for the new services and controller to ensure functionality and error handling.
