+++ backend/src/document-processor/services/aws-bedrock.service.spec.ts
